git-svn-id: http://svn.macosforge.org/repository/ruby/MacRuby/trunk@675 23306eb0-4c56-4727-a40e-e92c0eb68959

+++ bs.c
@@ -252,7 +252,7 @@ undecorate_struct_type(const char *src, char *dest, size_t dest_len,
     p_src = pos + 1;
     pos = strchr(p_src, '"');
     if (pos == NULL) {
-      fprintf(stderr, "Can't find the end of field delimiter starting at %d\n", p_src - src);
+      fprintf(stderr, "Can't find the end of field delimiter starting at %d\n", (int)(p_src - src));
       goto bails; 
     }
     if (field != NULL) {
@@ -307,7 +307,7 @@ undecorate_struct_type(const char *src, char *dest, size_t dest_len,
       }
 
       if (ok == false) {
-        fprintf(stderr, "Can't find the field encoding starting at %d\n", p_src - src);
+        fprintf(stderr, "Can't find the field encoding starting at %d\n", (int)(p_src - src));
         goto bails;
       }
